+++ platypush/__init__.py
@@ -24,7 +24,7 @@ from .utils import set_thread_name
 
 
 __author__ = 'Fabio Manganiello <info@fabiomanganiello.com>'
-__version__ = '0.20.5'
+__version__ = '0.20.6'
 
 logger = logging.getLogger('platypush')